pyproject.toml
[project]
name = "django-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.12",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.4.0",
    "django-cors-headers>=4.7.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.1",
    "supabase>=2.24.0",
]
